Create background.js; export to index.js
