+++ src/styles.css
@@ -41,10 +41,10 @@ body {
 }
 
 .theLogo:hover {
-	-webkit-filter: drop-shadow(1px 1px 0 rgb(155, 182, 0))
-                  drop-shadow(-1px -1px 0 rgb(155, 182, 0));
-  filter: drop-shadow(0px 0px 2px rgb(155, 182, 0)) 
-          drop-shadow(0px -1px 1px rgb(155, 182, 0));
+	-webkit-filter: drop-shadow(1px 1px 0 white)
+                  drop-shadow(-1px -1px 0 white);
+  filter: drop-shadow(0px -1px 1px white) 
+          drop-shadow(0px -1px 1px white);
 	cursor: pointer;
 }
